[*] model: force f32 in ModernBERT golden generators (fix token-parity bf16 drift)
The parity tests run Svod in f32, but the golden generators loaded the
reference model with from_pretrained() and no dtype, so a checkpoint whose
config defaults to bfloat16 would run in bf16 and produce bf16-rounded
logits. The NER fine-tune (sanketrai/...conll2003-english-ner) is saved in
bf16, so golden_token.safetensors was 100% bf16-exact and Svod's correct f32
output drifted by ~0.13 (token_logits_match_pytorch failed at [0][0]).

Force torch_dtype=torch.float32 in all three generators so a bf16-default
checkpoint can never leak into an f32-parity golden. The base and SST-2
checkpoints are f32-native, so their goldens were already f32 by luck.

Svod's token classifier was correct all along; no Rust changes. After
regenerating golden_token.safetensors in f32, token_logits_match_pytorch
passes at 5.7e-5 (was 0.13).

+++ model/scripts/generate_modernbert_classifier_golden.py
@@ -45,7 +45,10 @@ def main() -> None:
 
     device = "cpu"
     tokenizer = AutoTokenizer.from_pretrained(args.repo)
-    model = AutoModelForSequenceClassification.from_pretrained(args.repo).to(device).eval()
+    # Force f32: the parity tests run Svod in f32, so the golden must too.
+    # Without this a checkpoint whose config defaults to bfloat16 would load/run
+    # in bf16 and produce bf16-rounded logits that Svod's f32 output can't match.
+    model = AutoModelForSequenceClassification.from_pretrained(args.repo, torch_dtype=torch.float32).to(device).eval()
 
     enc = tokenizer(
         SAMPLE_SENTENCES,
